requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv==1.0.0
pytest==7.4.3
pytest-playwright==0.4.3
gunicorn==21.2.0
Flask-Limiter==3.5.0
requests==2.31.0
